Final structure

+++ src/main/java/com/example/dss/repository/DecisionRepository.java
@@ -3,6 +3,8 @@ package com.example.dss.repository;
 
 import com.example.dss.model.DecisionModel;
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.jpa.repository.Query;
+import org.springframework.data.repository.query.Param;
 
 import java.util.List;
 
@@ -11,5 +13,5 @@ public interface DecisionRepository extends JpaRepository<DecisionModel, Long> {
     List<DecisionModel> findAllDecisionById(int id);
 
     @Query("SELECT decision")
-    List<int> findDecisionIdsByUserId(@Param("userId") int userId);
+    List<Long> findDecisionIdsByUserId(@Param("userId") Long userId);
 }
